# conduit-api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Conduit API",
    "description": "REST API for the Conduit data pipeline orchestrator. Conduit is a Rust-native orchestration engine for DAGs (directed acyclic graphs) that compiles and schedules data pipelines with support for complex dependencies, dynamic scheduling, and schema contracts.",
    "version": "0.1.0",
    "contact": {
      "name": "Conduit Community"
    }
  },
  "servers": [
    {
      "url": "http://localhost:9000/api/v1",
      "description": "Local development server"
    },
    {
      "url": "https://api.conduit.dev/api/v1",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/dags/{dagId}/runs": {
      "get": {
        "tags": ["Runs"],
        "summary": "List runs for a DAG",
        "description": "Get all runs (executions) for a specific DAG.",
        "operationId": "listRuns",
        "parameters": [
          {
            "name": "dagId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50
            },
            "description": "Maximum number of runs to return"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["pending", "running", "success", "failed", "skipped"]
            },
            "description": "Filter by run status"
          }
        ],
        "responses": {
          "200": {
            "description": "List of runs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "runs": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/DagRun"
                      }
                    },
                    "total": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "tags": ["Runs"],
        "summary": "Trigger a DAG run",
        "description": "Queue a new execution of a DAG. Returns immediately with run ID.",
        "operationId": "triggerRun",
        "parameters": [
          {
            "name": "dagId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "logical_date": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Override the logical execution date"
                  },
                  "config": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "string"
                    },
                    "description": "Runtime configuration overrides"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Run queued successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "runId": {
                      "type": "string",
                      "example": "run_daily_etl_20250323_145030"
                    },
                    "dagId": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string",
                      "enum": ["queued", "pending", "running", "success", "failed"]
                    },
                    "taskStates": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "string"
                      }
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
    "/environments": {
      "get": {
        "tags": ["Environments"],
        "summary": "List environments",
        "description": "Get all virtual environments (dev, staging, production, etc.).",
        "operationId": "listEnvironments",
        "responses": {
          "200": {
            "description": "List of environments",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "environments": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Environment"
                      }
                    },
                    "total": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "tags": ["Environments"],
        "summary": "Create environment",
        "description": "Create a new virtual environment.",
        "operationId": "createEnvironment",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "staging"
                  },
                  "based_on": {
                    "type": "string",
                    "description": "Parent environment to clone from",
                    "example": "production"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Environment created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "snapshotCount": {
                      "type": "integer"
                    },
                    "basedOn": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
    "/apply": {
      "post": {
        "tags": ["Plan/Apply"],
        "summary": "Apply deployment plan",
        "description": "Apply a previously generated plan to update the target environment.",
        "operationId": "applyPlan",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "environment": {
                    "type": "string",
                    "default": "production"
                  },
                  "plan_id": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Plan applied",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "environment": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
    "/cluster/workers/{id}/drain": {
      "post": {
        "tags": ["Cluster"],
        "summary": "Drain worker",
        "description": "Gracefully drain a worker (stop accepting new tasks).",
        "operationId": "drainWorker",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Worker draining started",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "worker_id": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "ApiKey",
        "description": "API key authentication. Use format: Authorization: Bearer <api_key>"
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid authentication",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Insufficient permissions",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "DagRun": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "dagId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["pending", "running", "success", "failed", "skipped"]
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "endedAt": {
            "type": "string",
            "format": "date-time"
          },
          "taskStates": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "Map of task_id -> status"
          },
          "triggeredBy": {
            "type": "string",
            "enum": ["scheduler", "api", "manual"]
          }
        }
      },
      "Environment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "snapshotCount": {
            "type": "integer"
          },
          "basedOn": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}